Вынес объявление типов action в отдельный файл. Написал хуки useActions для каждого редьюсера
