+++ src/component/Converter/ExtensionChanger/ExtensionChangerContainer.js
@@ -2,7 +2,7 @@ import { connect } from "react-redux";
 import {
   changeExtensionInStoreAC,
   getExtFromServerTC,
-} from "../../../redux/fileUploaderReducer/fileUploaderReducer";
+} from "../../../redux/fileUploaderReducer/fileUploaderActions";
 
 import { ExtensionChanger } from "./ExtensionChanger";
 const mapStateToProps = (state) => {
