+++ src/redux/authReducer/authReducer.ts
@@ -1,11 +1,6 @@
-import { Dispatch } from 'redux';
-import { api } from '../../api/api';
-import { IAction } from '../interface';
 import { IAuthReducerState, AuthActionTypes, TAuthReducerActions} from './interface/index';
 
-const SET_IS_AUTH = "SET-IS-AUTH";
-const LOG_OUT = "LOG-OUT";
-const SET_BED_REQ = "SET-BED-REQ";
+
 const initialState:IAuthReducerState = { bedReq: null, isAuth: localStorage.getItem("token") };
 export const authReducer = (state = initialState, action:TAuthReducerActions):IAuthReducerState=> {
   switch (action.type) {
@@ -29,38 +24,4 @@ export const authReducer = (state = initialState, action:TAuthReducerActions):IA
       return state;
   }
 };
-export const setIsAuthAC= (isAuth:string|null) => ({ type: SET_IS_AUTH, isAuth });
-export const setBedReqAC = (errorMessage:string|null) => ({
-  type: SET_BED_REQ,
-  bedReq: errorMessage,
-});
 
-export const logOut = () => ({ type: LOG_OUT });
-export const logIn = (email:string, password:string) => {
-  return async (dispatch:Dispatch<IAction>) => {
-    try {
-      let response = await api.post("Authenticate/authenticate", {
-        email,
-        password,
-      });
-      localStorage.setItem("token", response.data.token);
-      dispatch(setIsAuthAC(localStorage.getItem("token")));
-    } catch (e:any) {
-      dispatch(setBedReqAC(e.response.data.message));
-    }
-  };
-};
-export const regUser = (email:string, password:string) => {
-  return async (dispatch:Dispatch<IAction>) => {
-    try {
-      let response = await api.post("Authenticate/register", {
-        email,
-        password,
-      });
-      dispatch(setBedReqAC(localStorage.getItem("token")));
-      localStorage.setItem("token", response.data.token);
-    } catch (e:any) {
-      dispatch(setBedReqAC(e.response.data.message));
-    }
-  };
-};
