+++ src/redux/fileUploaderReducer/fileUploaderReducer.ts
@@ -1,14 +1,5 @@
-import { Dispatch } from "redux";
-import { fileUploaderAPI } from "../../api/fileUploaderAPI";
-import { initialAPI } from "../../api/initialAPI";
-import { IAction } from "../interface";
 import { TFileUploaderA, IImageHistoryReducerState, fileUploaderAE } from "./interface/index";
-const ON_FILE_CHANGE = "ON-FILE-CHANGE";
-const ON_FILE_UPLOAD = "ON-FILE-UPLOAD";
-const CHANGE_EXTENSION_IN_STORE = "CHANGE-EXTENSION-IN-STORE";
-const LOAD_EXT = "LOAD-EXT";
-const UPLOAD_BUTTON_DISABLE = "UPLOAD-BUTTON-DISABLE";
-const SET_IMAGE_FILE_NAME = "SET-IMAGE-FILE-NAME";
+
 
 let initialState:IImageHistoryReducerState = {
   imageFile: null,
@@ -63,49 +54,5 @@ const fileUploaderReducer = (
     }
   }
 };
-export const changeExtensionInStoreAC = (ext:string) => ({
-  type: CHANGE_EXTENSION_IN_STORE,
-  extensionTo: ext,
-});
-export const onFileChangeAC = (imageFile:File) => ({
-  type: ON_FILE_CHANGE,
-  imageFile: imageFile,
-});
-export const onFileUploadAC = (URLconvertedFile:string) => ({
-  type: ON_FILE_UPLOAD,
-  URLconvertedFile: URLconvertedFile,
-});
+
 export { fileUploaderReducer };
-export const loadExtAC = (extArr:string[]) => ({ type: LOAD_EXT, extArr: extArr });
-const uploadButtonDisablerAC = (boolean:boolean) => ({
-  type: UPLOAD_BUTTON_DISABLE,
-  boolean,
-});
-export const sendImageToServerTC = (imageFile:File, extensionTo:string) => {
-  return async (dispatch:Dispatch<IAction>) => {
-    let formData = new FormData();
-    formData.append("ImageFile", imageFile);
-    formData.append("ExtensionTo", extensionTo);
-    dispatch(uploadButtonDisablerAC(true));
-    let URLconvertedFile = await fileUploaderAPI
-      .sendImageToServer(formData)
-      .then((response) => {
-        let URLconvertedFile = URL.createObjectURL(response.data);
-        return URLconvertedFile;
-      });
-    dispatch(setImageFileNameAC());
-    dispatch(uploadButtonDisablerAC(false));
-    dispatch(onFileUploadAC(URLconvertedFile));
-  };
-};
-
-export const getExtFromServerTC = () => {
-  return async (dispatch:Dispatch<IAction>) => {
-    let extArr = await initialAPI.initialExt().then((resp) => resp.data);
-    dispatch(loadExtAC(extArr));
-  };
-};
-
-const setImageFileNameAC = () => ({
-  type: SET_IMAGE_FILE_NAME,
-});
