+++ src/redux/hooks/useAppSelector.ts
@@ -1,4 +1,4 @@
 import { TypedUseSelectorHook, useSelector } from "react-redux";
-import { RootState } from "../redux/store";
+import { RootState } from "../store";
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
