+++ src/redux/hooks/useImageHistoryActions.ts
@@ -1,0 +1,7 @@
+import { bindActionCreators } from 'redux';
+import { useAppDispatch } from "./useAppDispatch"
+import * as imageHistoryActions from '../imageHistoryReducer/imageHistoryActions'
+export const useImageHistoryActions = ()=>{
+    const dispatch= useAppDispatch()
+    return bindActionCreators(imageHistoryActions,dispatch)
+}
